updata /root 2022-11-03

+++ root/gitbk.sh
@@ -4,12 +4,14 @@ obj="root"
 
 cd /root/Git/UnixSystemProgramming/
 
-git rm -r $obj
-git commit -m "remove old /root $(date +%F)"
-rm -rf $obj
+if [ -d "$obj" ]; then
+	git rm -r $obj
+	git commit -m "remove old /root $(date +%F)"
+	rm -rf $obj
+fi
 
 cp -r /root ./
 git add $obj
-git commit "updata /root $(date +%F)"
+git commit -m "updata /root $(date +%F)"
 
 git push
